requirements.txt
PyGithub==1.54.1
PyJWT==1.7.1
requests==2.25.1

